new version2.0.0

+++ packages/bitcore-wallet-service/stop.sh
@@ -1,12 +1,25 @@
 #!/bin/bash
 
+if [ "x${BITCORE_PATH}" == "x" ]; then
+  BITCORE_PATH=/root/bitcore
+fi
+
+MODULE_PATH=$BITCORE_PATH/packages
+
+cd $MODULE_PATH/bitcore-wallet-service
+
+
 stop_program ()
 {
   pidfile=$1
 
-  echo "Stopping Process - $pidfile. PID=$(cat $pidfile)"
-  kill -9 $(cat $pidfile)
-  rm $pidfile
+  if [ -f $pidfile ]; then
+    echo "Stopping Process - $pidfile. PID=$(cat $pidfile)"
+    kill -9 $(cat $pidfile)
+    rm -f $pidfile
+  else
+    echo "Stopping Process - $pidfile."
+  fi
   
 }
